requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
requests==2.32.3
urllib3==2.2.2
aiogram~=3.18.0

pyTelegramBotAPI~=4.22.0
APScheduler~=3.11.0
mysql-connector-python